# swagger.yaml
openapi: 3.0.0
info:
  title: Togather Backend
  description: togather backend development
  contact: {}
  version: "1.0"
servers:
  - url: https://7f45-2001-448a-4020-34d6-d082-8b76-9b90-9059.ap.ngrok.io/api/v1
    variables: {}
paths:
  /users:
    get:
      tags:
        - Users
      summary: Get All Users
      operationId: GetAllUsers
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
        - Users
      summary: Get Single User
      operationId: GetSingleUser
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Users
      summary: Delete User
      operationId: DeleteUser
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Users
      summary: Update User
      operationId: UpdateUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateUserRequest"
                - example:
                    mobile: "085123456780"
                    userName: aligator
            example:
              mobile: "085123456780"
              userName: aligator
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register User
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/RegisterUserRequest"
                - example:
                    name: Greta
                    email: testing2@gmail.com
                    mobile: "085123456789"
                    userName: testing2
                    password: "1234567"
            example:
              name: Greta
              email: testing2@gmail.com
              mobile: "085123456789"
              userName: testing2
              password: "1234567"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login User
      operationId: LoginUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/LoginUserRequest"
                - example:
                    email: joe35@gmail.com
                    password: password1234
            example:
              email: joe35@gmail.com
              password: password1234
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/login/google:
    post:
      tags:
        - Auth
      summary: Google Login
      operationId: GoogleLogin
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/GoogleLoginRequest"
                - example:
                    credential: eyJhbGciOiJSUzI1NiIsImtpZCI6IjI1NWNjYTZlYzI4MTA2MDJkODBiZWM4OWU0NTZjNDQ5NWQ3NDE4YmIiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2Nzg0NzEyMjAsImF1ZCI6IjU2MTI2NjE1NTYwMi1ubzMzazJ0NDJ1NTdtbXVsYWxhcTJzMTFpZGxhbzU3dC5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjEwNTEwOTYzMTU2MzkwNjA0Mjk2NyIsImVtYWlsIjoiYWRpd3NudUBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiNTYxMjY2MTU1NjAyLW5vMzNrMnQ0MnU1N21tdWxhbGFxMnMxMWlkbGFvNTd0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwibmFtZSI6IkFkaSBXaXNudSIsInBpY3R1cmUiOiJodHRwczovL2xoMy5nb29nbGV1c2VyY29udGVudC5jb20vYS9BR05teXhiT2ltZ1dxMXptcFBzaGdlWVF0X0N2cGwwUW9zRXhJbmo2LU1SYj1zOTYtYyIsImdpdmVuX25hbWUiOiJBZGkiLCJmYW1pbHlfbmFtZSI6Ildpc251IiwiaWF0IjoxNjc4NDcxNTIwLCJleHAiOjE2Nzg0NzUxMjAsImp0aSI6IjM2NTg2Y2MwOTU2ZmQyM2MwOTU2NmVhMjZjMTViOGUyZDYxMTZmOWYifQ.QGxnBB9XtZhnHuMQuYO-G2g4az8r8nbfBR9xl9RzjOZ_PT6wsT5C_ViytHFpb5vzxhjed-b1PmXeGjf_J0pOvvl6DA6JZrZHt6JQK0QTW1pHjTTYITzCTy3VIm2ifWA5RBCWmGqYh3gUbixrVf0zptuGDYsyOQQX4ow3LXQFwqgelOecWBNTnFTBVi7Gznh-cPwIYfuGwQtGf6yE1wbIFm3kcuDnO2yqLvZ9nx-TruZwfinYbp5FQFWtc54jIov8bTPE8bF9_E_xrYzpx0K0L9gWsOpYqBvPUnnWCFM3WFyXug87UvGAT8PfIhhDLQUjFcNscFpdGE6c-iweHBzIwQ
            example:
              credential: eyJhbGciOiJSUzI1NiIsImtpZCI6IjI1NWNjYTZlYzI4MTA2MDJkODBiZWM4OWU0NTZjNDQ5NWQ3NDE4YmIiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2Nzg0NzEyMjAsImF1ZCI6IjU2MTI2NjE1NTYwMi1ubzMzazJ0NDJ1NTdtbXVsYWxhcTJzMTFpZGxhbzU3dC5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjEwNTEwOTYzMTU2MzkwNjA0Mjk2NyIsImVtYWlsIjoiYWRpd3NudUBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiNTYxMjY2MTU1NjAyLW5vMzNrMnQ0MnU1N21tdWxhbGFxMnMxMWlkbGFvNTd0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwibmFtZSI6IkFkaSBXaXNudSIsInBpY3R1cmUiOiJodHRwczovL2xoMy5nb29nbGV1c2VyY29udGVudC5jb20vYS9BR05teXhiT2ltZ1dxMXptcFBzaGdlWVF0X0N2cGwwUW9zRXhJbmo2LU1SYj1zOTYtYyIsImdpdmVuX25hbWUiOiJBZGkiLCJmYW1pbHlfbmFtZSI6Ildpc251IiwiaWF0IjoxNjc4NDcxNTIwLCJleHAiOjE2Nzg0NzUxMjAsImp0aSI6IjM2NTg2Y2MwOTU2ZmQyM2MwOTU2NmVhMjZjMTViOGUyZDYxMTZmOWYifQ.QGxnBB9XtZhnHuMQuYO-G2g4az8r8nbfBR9xl9RzjOZ_PT6wsT5C_ViytHFpb5vzxhjed-b1PmXeGjf_J0pOvvl6DA6JZrZHt6JQK0QTW1pHjTTYITzCTy3VIm2ifWA5RBCWmGqYh3gUbixrVf0zptuGDYsyOQQX4ow3LXQFwqgelOecWBNTnFTBVi7Gznh-cPwIYfuGwQtGf6yE1wbIFm3kcuDnO2yqLvZ9nx-TruZwfinYbp5FQFWtc54jIov8bTPE8bF9_E_xrYzpx0K0L9gWsOpYqBvPUnnWCFM3WFyXug87UvGAT8PfIhhDLQUjFcNscFpdGE6c-iweHBzIwQ
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /rooms:
    get:
      tags:
        - Rooms
      summary: Get All Rooms
      operationId: GetAllRooms
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /rooms/auth:
    post:
      tags:
        - Rooms
      summary: Create Room
      operationId: CreateRoom
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateRoomRequest"
                - example:
                    roomName: testing status 12
                    membersCount: 3
                    members:
                      - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
                      - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
                      - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
            example:
              roomName: testing status 12
              membersCount: 3
              members:
                - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
                - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
                - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Rooms
      summary: Get User Created Rooms
      operationId: GetUserCreatedRooms
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /rooms/auth/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The room ID
    get:
      tags:
        - Rooms
      summary: Get Single Room
      operationId: GetSingleRoom
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Rooms
      summary: Edit Room
      operationId: EditRoom
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/EditRoomRequest"
                - example:
                    members:
                      - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
                      - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
                      - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
                      - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
            example:
              members:
                - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
                - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
                - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
                - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Rooms
      summary: Delete Room
      operationId: DeleteRoom
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /rooms/auth/{id}/join:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The room ID
    patch:
      tags:
        - Rooms
      summary: Join room
      operationId: Joinroom
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /rooms/auth/{id}/leave:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The room ID
    patch:
      tags:
        - Rooms
      summary: Leave Room
      operationId: LeaveRoom
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
  schemas:
    CreateUserRequest:
      title: CreateUserRequest
      required:
        - name
        - email
        - mobile
        - userName
        - address
        - age
        - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        userName:
          type: string
        address:
          type: string
        age:
          type: string
        password:
          type: string
      example:
        name: Joe
        email: johne3@gmail.com
        mobile: "085123456789"
        userName: joe232345
        address: solo
        age: "30"
        password: password1234
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
        - mobile
        - userName
      type: object
      properties:
        mobile:
          type: string
        userName:
          type: string
      example:
        mobile: "085123456780"
        userName: aligator
    RegisterUserRequest:
      title: RegisterUserRequest
      required:
        - name
        - email
        - mobile
        - userName
        - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        userName:
          type: string
        password:
          type: string
      example:
        name: Greta
        email: testing2@gmail.com
        mobile: "085123456789"
        userName: testing2
        password: "1234567"
    LoginUserRequest:
      title: LoginUserRequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: joe35@gmail.com
        password: password1234
    GoogleLoginRequest:
      title: GoogleLoginRequest
      required:
        - credential
      type: object
      properties:
        credential:
          type: string
      example:
        credential: eyJhbGciOiJSUzI1NiIsImtpZCI6IjI1NWNjYTZlYzI4MTA2MDJkODBiZWM4OWU0NTZjNDQ5NWQ3NDE4YmIiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2Nzg0NzEyMjAsImF1ZCI6IjU2MTI2NjE1NTYwMi1ubzMzazJ0NDJ1NTdtbXVsYWxhcTJzMTFpZGxhbzU3dC5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjEwNTEwOTYzMTU2MzkwNjA0Mjk2NyIsImVtYWlsIjoiYWRpd3NudUBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiNTYxMjY2MTU1NjAyLW5vMzNrMnQ0MnU1N21tdWxbGFxMnMxMWlkbGFvNTd0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwibmFtZSI6IkFkaSBXaXNudSIsInBpY3R1cmUiOiJodHRwczovL2xoMy5nb29nbGV1c2VyY29udGVudC5jb20vYS9BR05teXhiT2ltZ1dxMXptcFBzaGdlWVF0X0N2cGwwUW9zRXhJbmo2LU1SYj1zOTYtYyIsImdpdmVuX25hbWUiOiJBZGkiLCJmYW1pbHlfbmFtZSI6Ildpc251IiwiaWF0IjoxNjc4NDcxNTIwLCJleHAiOjE2Nzg0NzUxMjAsImp0aSI6IjM2NTg2Y2MwOTU2ZmQyM2MwOTU2NmVhMjZjMTViOGUyZDYxMTZmOWYifQ.QGxnBB9XtZhnHuMQuYO-G2g4az8r8nbfBR9xl9RzjOZ_PT6wsT5C_ViytHFpb5vzxhjed-b1PmXeGjf_J0pOvvl6DA6JZrZHt6JQK0QTW1pHjTTYITzCTy3VIm2ifWA5RBCWmGqYh3gUbixrVf0zptuGDYsyOQQX4ow3LXQFwqgelOecWBNTnFTBVi7Gznh-cPwIYfuGwQtGf6yE1wbIFm3kcuDnO2yqLvZ9nx-TruZwfinYbp5FQFWtc54jIov8bTPE8bF9_E_xrYzpx0K0L9gWsOpYqBvPUnnWCFM3WFyXug87UvGAT8PfIhhDLQUjFcNscFpdGE6c-iweHBzIwQ
    CreateRoomRequest:
      title: CreateRoomRequest
      required:
        - roomName
        - membersCount
        - members
      type: object
      properties:
        roomName:
          type: string
        membersCount:
          type: integer
          format: int32
        members:
          type: array
          items:
            type: string
          description: ""
      example:
        roomName: testing status 12
        membersCount: 3
        members:
          - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
          - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
          - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
    EditRoomRequest:
      title: EditRoomRequest
      required:
        - members
      type: object
      properties:
        members:
          type: array
          items:
            type: string
          description: ""
      example:
        members:
          - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
          - nd1DQiIE8B1uFEea-oGRDqXI0LfLBW_2lHlW
          - BNSSivCS7Wb-oHuMoJq0TeDikZP6I2bCWWQW
          - hdbNuN48saKMUdxrkilz2dbyrCBy8IJS8vNT
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
tags:
  - name: Users
    description: ""
  - name: Auth
    description: ""
  - name: Rooms
    description: ""